+++ src/main/java/nl/ramsolutions/sw/magik/jacoco/generators/BaseReportGenerator.java
@@ -10,6 +10,7 @@ import org.jacoco.core.tools.ExecFileLoader;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Path;
 
 /**
  * Base report generator.
@@ -19,7 +20,7 @@ public abstract class BaseReportGenerator {
     private final File productDirectory;
     private final File outputFile;
     private final File executionDataFile;
-    private final boolean filterPrimaryClasses;
+    private final boolean filterExecutableClasses;
     private final ExecFileLoader execFileLoader = new ExecFileLoader();
     private Sw5LibReader libReader;
 
@@ -33,17 +34,17 @@ public abstract class BaseReportGenerator {
      * @param productDirectory File to product directory.
      * @param executionDataFile File to {@literal jacoco.exec}.
      * @param outputFile File to report directory.
-     * @param filterPrimaryClasses Filter primary classes.
+     * @param filterExecutableClasses Filter executable classes.
      */
     protected BaseReportGenerator(
             final File productDirectory,
             final File executionDataFile,
             final File outputFile,
-            final boolean filterPrimaryClasses) {
+            final boolean filterExecutableClasses) {
         this.productDirectory = productDirectory;
         this.executionDataFile = executionDataFile;
         this.outputFile = outputFile;
-        this.filterPrimaryClasses = filterPrimaryClasses;
+        this.filterExecutableClasses = filterExecutableClasses;
     }
 
     protected File getOutputFile() {
@@ -88,12 +89,14 @@ public abstract class BaseReportGenerator {
         final Analyzer analyzer = new Analyzer(dataStore, coverageBuilder);
         final File libsDirectory = new File(productDirectory, "libs");
         analyzer.analyzeAll(libsDirectory);
-        final String title = this.productDirectory.getName();
-        final IBundleCoverage bundleCoverage = coverageBuilder.getBundle(title);
+        final String name = this.productDirectory.getName();
+        final IBundleCoverage bundleCoverage = coverageBuilder.getBundle(name);
 
-        // Merge method coverages (Magik), filter primary classes if needed.
-        final MagikBundleCoverageConverter bundleCoverageConverter = new MagikBundleCoverageConverter(this.libReader);
-        return bundleCoverageConverter.convert(bundleCoverage, this.filterPrimaryClasses);
+        // Merge method coverages (Magik), filter executable classes if needed.
+        final MagikBundleCoverageConverter bundleCoverageConverter =
+            new MagikBundleCoverageConverter(this.libReader, bundleCoverage, this.filterExecutableClasses);
+        final IBundleCoverage newBundleCoverage = bundleCoverageConverter.convert();
+        return newBundleCoverage;
     }
 
     private void loadExecutionData() throws IOException {
@@ -101,7 +104,8 @@ public abstract class BaseReportGenerator {
     }
 
     private void loadSw5Libs() throws IOException {
-        this.libReader = new Sw5LibReader(this.productDirectory.toPath());
+        final Path productPath = this.productDirectory.toPath();
+        this.libReader = new Sw5LibReader(productPath);
     }
 
 }