+++ src/main/java/nl/ramsolutions/sw/magik/jacoco/generators/MagikNames.java
@@ -1,22 +1,17 @@
 package nl.ramsolutions.sw.magik.jacoco.generators;
 
-import nl.ramsolutions.sw.magik.jacoco.sw5lib.Sw5LibAnalyzer;
 import nl.ramsolutions.sw.magik.jacoco.sw5lib.Sw5LibReader;
 import org.jacoco.report.ILanguageNames;
 import org.jacoco.report.JavaNames;
-
-import java.util.Map;
 
 /**
  * Magik names extractor for JaCoCo report generator.
  */
 public class MagikNames implements ILanguageNames {
 
-    private final Sw5LibAnalyzer libAnalyzer;
     private final JavaNames javaNames;
 
     public MagikNames(final Sw5LibReader libReader) {
-        this.libAnalyzer = new Sw5LibAnalyzer(libReader);
         this.javaNames = new JavaNames();
     }
 
@@ -45,15 +40,7 @@ public class MagikNames implements ILanguageNames {
             final String vmmethodname,
             final String vmdesc,
             final String vmsignature) {
-        // Perhaps we would want some caching...
-        Map<String, String> methodNameMapping = this.libAnalyzer.extractAllMethodNames();
-        String javaMethodName = this.libAnalyzer.keyForClassMethodName(vmclassname, vmmethodname);
-        final String methodName = methodNameMapping.get(javaMethodName);
-        if (methodName != null) {
-            return methodName;
-        }
-
-        return this.javaNames.getMethodName(vmclassname, vmmethodname, vmdesc, vmsignature);
+        return vmmethodname;
     }
 
     @Override
@@ -64,6 +51,10 @@ public class MagikNames implements ILanguageNames {
             final String vmsignature) {
         final String qualifiedClassName = this.getQualifiedClassName(vmclassname);
         final String methodName = this.getMethodName(vmclassname, vmmethodname, vmdesc, vmsignature);
+        if (methodName.startsWith("[")) {
+            return String.format("%s%s", qualifiedClassName, methodName);
+        }
+
         return String.format("%s.%s", qualifiedClassName, methodName);
     }