+++ src/test/java/nl/ramsolutions/sw/magik/jacoco/sw5lib/Sw5LibAnalyzerTest.java
@@ -4,7 +4,6 @@ import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
 import java.nio.file.Path;
-import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -23,23 +22,31 @@ class Sw5LibAnalyzerTest {
     void test() throws IOException {
         final Sw5LibReader libReader = Sw5LibAnalyzerTest.getLibReader();
         final Sw5LibAnalyzer libAnalyzer = new Sw5LibAnalyzer(libReader);
-        final Map<String, String> extractAllMethodNames = libAnalyzer.extractAllMethodNames();
-        assertThat(extractAllMethodNames).containsOnly(
-            Map.entry(
-                "magik.fixture_product.fixture_module.char16_vector_32.char16_vector__method1",
-                "char16_vector.method1()"),
-            Map.entry(
-                "magik.fixture_product.fixture_module.char16_vector_32.char16_vector__method2",
-                "char16_vector.method2()"),
-            Map.entry(
-                "magik.fixture_product.fixture_module.char16_vector_32.char16_vector__method3?",
-                "char16_vector.method3?()"),
-            Map.entry(
-                "magik.fixture_product.fixture_module.other_39.float__plus_100",
-                "float.plus_100()"),
-            Map.entry(
-                "magik.fixture_product.fixture_module.other_39.integer__plus_100",
-                "integer.plus_100()"));
+
+        final String magikMethod1 = libAnalyzer.getMagikMethodName(
+            "magik.fixture_product.fixture_module.char16_vector_32",
+            "char16_vector__method1");
+        assertThat(magikMethod1).isEqualTo("char16_vector.method1()");
+
+        final String magikMethod2 = libAnalyzer.getMagikMethodName(
+            "magik.fixture_product.fixture_module.char16_vector_32",
+            "char16_vector__method2");
+        assertThat(magikMethod2).isEqualTo("char16_vector.method2()");
+
+        final String magikMethod3 = libAnalyzer.getMagikMethodName(
+            "magik.fixture_product.fixture_module.char16_vector_32",
+            "char16_vector__method3?");
+        assertThat(magikMethod3).isEqualTo("char16_vector.method3?()");
+
+        final String magikMethod4 = libAnalyzer.getMagikMethodName(
+            "magik.fixture_product.fixture_module.other_39",
+            "float__plus_100");
+        assertThat(magikMethod4).isEqualTo("float.plus_100()");
+
+        final String magikMethod5 = libAnalyzer.getMagikMethodName(
+            "magik.fixture_product.fixture_module.other_39",
+            "integer__plus_100");
+        assertThat(magikMethod5).isEqualTo("integer.plus_100()");
     }
 
 }